Code clean up, remove dupes, add xml comments, use long for streaminfo

+++ SevenZip/EventArguments/NewSourceStreamEventArgs.cs
@@ -1,15 +1,27 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace SevenZip.EventArguments
 {
+    /// <summary>
+    /// Provides data for the NewSourceStream event.
+    /// </summary>
     public class NewSourceStreamEventArgs : EventArgs
     {
-        public readonly uint StreamIndex;
+        /// <summary>
+        /// Gets the index of the stream.
+        /// </summary>
+        public readonly long StreamIndex;
+
+        /// <summary>
+        /// Gets or sets the source request object.
+        /// </summary>
         public Object SourceRequest;
 
-        public NewSourceStreamEventArgs(uint streamIndex)
+        /// <summary>
+        /// Initializes a new instance of the NewSourceStreamEventArgs class.
+        /// </summary>
+        /// <param name="streamIndex">The index of the stream.</param>
+        public NewSourceStreamEventArgs(long streamIndex)
         {
             this.StreamIndex = streamIndex;
         }
